+++ docxbox/docx/renderer/contentType/docx_renderer_list.h
@@ -26,6 +26,8 @@ class docx_renderer_list: docx_renderer {
   std::string Render(bool is_ordered);
   std::string Render() override;
 
+  std::string generic_root_tag_;
+
  private:
   bool is_ordered_ = false;
   std::vector<std::string> items_;