+++ docxbox/docx/renderer/docx_renderer_delegate.h
@@ -10,6 +10,7 @@
 #include <docxbox/docx/renderer/contentType/docx_renderer_image.h>
 #include <docxbox/docx/renderer/contentType/docx_renderer_link.h>
 #include <docxbox/docx/renderer/contentType/docx_renderer_list.h>
+#include <docxbox/docx/renderer/contentType/docx_renderer_paragraph.h>
 #include <docxbox/docx/renderer/contentType/docx_renderer_table.h>
 #include <docxbox/docx/renderer/contentType/docx_renderer_text.h>
 
@@ -25,8 +26,7 @@ class docx_renderer_delegate {
 
   std::string path_extract_;
 
-  // TODO(kay): detect 1st child tag generic / spare manual specification
-  std::string replacement_xml_first_child_tag_ = "w:r";
+  std::string replacement_xml_root_tag_ = "w:r";
 
   // @see docx_xml_replace::SetImageRelationshipId()
   std::string image_relationship_id_ = "";
@@ -39,14 +39,13 @@ class docx_renderer_delegate {
       int level, const std::string &json, std::string &markup);
 
   std::string &RenderHyperlink(const std::string &json, std::string &markup);
+  std::string &RenderImage(const std::string &json, std::string &markup);
 
   std::string &RenderList(
       bool is_ordered, const std::string &json, std::string &markup);
 
-  std::string &RenderImage(const std::string &json, std::string &markup);
-
+  std::string &RenderParagraph(const std::string &json, std::string &markup);
   std::string &RenderTable(const std::string &json, std::string &markup);
-
   std::string &RenderText(const std::string &json, std::string &markup);
 };